modularizing code and components. adding directories and modules where necessary. Fixing imports and separating css files
