+++ src/App.jsx
@@ -1,18 +1,12 @@
 import { useEffect, useState } from "react"
-import stevePic from "./assets/steve.png"
+import { AddJoke } from "./components/jokes/AddJoke"
+import { Header } from "./components/header/Header"
+import { JokesList } from "./components/jokes/JokesList"
+import { getAllJokes } from "./services/jokeService"
 import "./App.css"
-import {
-	deleteJoke,
-	editJoke,
-	getAllJokes,
-	postNewJoke,
-} from "./services/jokeService"
 
 export const App = () => {
-	const [newJoke, setNewJoke] = useState("")
 	const [allJokes, setAllJokes] = useState([])
-	const [untoldJokes, setUntoldJokes] = useState([])
-	const [toldJokes, setToldJokes] = useState([])
 
 	const getAndSetAllJokes = () => {
 		getAllJokes().then((jokesArray) => setAllJokes(jokesArray))
@@ -21,115 +15,11 @@ export const App = () => {
 		getAndSetAllJokes()
 	}, [])
 
-	const getAllToldJokes = () => {
-		const filteredToldJokes = allJokes.filter((joke) => joke.told)
-		setToldJokes(filteredToldJokes)
-	}
-
-	const getAllUntoldJokes = () => {
-		const filteredUntoldJokes = allJokes.filter((joke) => joke.told === false)
-		setUntoldJokes(filteredUntoldJokes)
-	}
-
-	useEffect(() => {
-		getAllToldJokes()
-		getAllUntoldJokes()
-	}, [allJokes])
-
-	const addNewJoke = () => {
-		if (newJoke) {
-			postNewJoke(newJoke).then(setNewJoke("")).then(getAndSetAllJokes)
-		}
-	}
-	const handleToggleJoke = async (joke) => {
-		joke.told = !joke.told
-		await editJoke(joke)
-		getAndSetAllJokes()
-	}
-
-	const handleDeleteJoke = async (joke) => {
-		await deleteJoke(joke)
-		getAndSetAllJokes()
-	}
-
 	return (
 		<div className="app-container">
-			<div className="app-heading">
-				<div className="app-heading-circle">
-					<img className="app-logo" src={stevePic} alt="Good job Steve" />
-				</div>
-				<h1 className="app-heading-text">Chuckle Checklist</h1>
-			</div>
-			<h2>Add Joke</h2>
-			<div className="joke-add-form">
-				<input
-					className="joke-input"
-					type="text"
-					value={newJoke}
-					placeholder="New One Liner"
-					onChange={(event) => {
-						setNewJoke(event.target.value)
-					}}
-				/>
-				<button className="joke-input-submit" onClick={addNewJoke}>
-					Add
-				</button>
-			</div>
-			<div className="joke-lists-container">
-				<div className="joke-list-container">
-					<h2>
-						Told Jokes <span className="told-count">{toldJokes.length}</span>
-					</h2>
-					{toldJokes.map((jokeObject) => {
-						return (
-							<li className="joke-list-item" key={jokeObject.id}>
-								<p className="joke-list-item-text">{jokeObject.text}</p>
-								<button
-									className="joke-list-action-toggle"
-									onClick={async () => {
-										handleToggleJoke(jokeObject)
-									}}>
-									<i className="fa-regular fa-face-laugh-squint" />
-								</button>
-								<button
-									className="joke-list-action-delete"
-									onClick={async () => {
-										handleDeleteJoke(jokeObject)
-									}}>
-									<i className="fa-solid fa-trash"></i>
-								</button>
-							</li>
-						)
-					})}
-				</div>
-				<div className="joke-list-container">
-					<h2>
-						Untold Jokes{" "}
-						<span className="untold-count">{untoldJokes.length}</span>
-					</h2>
-					{untoldJokes.map((jokeObject) => {
-						return (
-							<li className="joke-list-item" key={jokeObject.id}>
-								<p className="joke-list-item-text">{jokeObject.text}</p>
-								<button
-									className="joke-list-action-toggle"
-									onClick={async () => {
-										handleToggleJoke(jokeObject)
-									}}>
-									<i className="fa-regular fa-face-meh" />
-								</button>
-								<button
-									className="joke-list-action-delete"
-									onClick={async () => {
-										handleDeleteJoke(jokeObject)
-									}}>
-									<i className="fa-solid fa-trash"></i>
-								</button>
-							</li>
-						)
-					})}
-				</div>
-			</div>
+			<Header />
+			<AddJoke getAndSetAllJokes={getAndSetAllJokes} />
+			<JokesList allJokes={allJokes} getAndSetAllJokes={getAndSetAllJokes} />
 		</div>
 	)
 }
