+++ src/components/buttons/ChangeJokeStatus.css
@@ -1,0 +1,24 @@
+/** Classes for the action buttons. Wrap the buttons in their own div*/
+/** The toggle class is for the buttons to mark a joke as told or untold*/
+.joke-list-action-delete,
+.joke-list-action-toggle {
+	border: none;
+	padding: 5px;
+	margin-left: 0.25rem;
+	background: transparent;
+	color: var(--col-4);
+}
+
+/** Adds styling for when the user hovers over the delete button*/
+.joke-list-action-delete :hover {
+	color: var(--col-0);
+	cursor: pointer;
+	transition: all 0.1s ease-in;
+}
+
+/** Adds styling for when the user hovers over one of the toggle buttons*/
+.joke-list-action-toggle :hover {
+	color: var(--col-5);
+	cursor: pointer;
+	transition: all 0.1s ease-in;
+}
